// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/phrase-context-groups-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PhraseFullContextGroupDto } from '../models/phrase-full-context-group-dto';
import { PhraseResolvedQueryDto } from '../models/phrase-resolved-query-dto';
import { PhraseSelectedPathDto } from '../models/phrase-selected-path-dto';
export interface PhraseContextGroupsResponse {
  activeBuildId: string;
  followingSelection: PhraseSelectedPathDto;
  items: Array<PhraseFullContextGroupDto>;
  nextCursor: string | null;
  previousSelection: PhraseSelectedPathDto;
  query: PhraseResolvedQueryDto;
  totalCount: number;
}