// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/phrase-context-occurrences-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PhraseContextOccurrenceDto } from '../models/phrase-context-occurrence-dto';
import { PhraseFullContextDto } from '../models/phrase-full-context-dto';
export interface PhraseContextOccurrencesResponse {
  activeBuildId: string;
  context: PhraseFullContextDto;
  items: Array<PhraseContextOccurrenceDto>;
  nextCursor: string | null;
  totalCount: number;
}